requirements.txt
discord-py==1.4.1
praw==7.1.0
python==3.8.5
python-dotenv==0.14.0